requirements.txt
sqlalchemy==2.0.40
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic==2.11.1
tronpy==0.5.0
python-dotenv